Solution to Check Even or Odd

+++ Check Even or Odd@@6788dd63490637cfc44e1b8d/code.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 int main()
 {
-    int a;
+    double a;
     if (a%2==0)
     printf("Even");
     else printf("Odd");
